fixes race condition that was causing problems establishing P2P connections
the fix is kind of buried in all the refactoring, but the problem was
when a watcher tried to join a game before they got their own Peer ID,
it would silently fail.  in other words, after creating a `new Peer()`
instance, you have to wait for the 'open' event to fire on it before
trying to use `peer.connect(otherPeerId)`.

i'm so confident in my fix that i'm removing the "cheap, unreliable
software" disclaimer.